Time replacement of description appropriately
This handles two scenarios for when to update the description element:
- When the element already exists
- When the element will soon exist

In this commit, when the element already exists, it is updated immediately. When it doesn't, however, the observer now waits for the change in the DOM to happen, in which it then updates (as the first change it records has the element, from my own testing). The purpose in this was to fix a common issue where the description would sometimes not be updated. The previous attempt of trying to wait a second was flawed since it didn't actually resolve the issue (which was waiting for a change to be observed when it already happened).